FAGNN_optimized/requirements.txt
tqdm
numpy
pandas
cupy-cuda12x            # GPU-accelerated drop‑in replacement for NumPy
cudf-cu12x              # GPU DataFrame library (cpandas)
numba                   # JIT compiler for Python & NumPy
pyarrow
seaborn
jupyter
imblearn
ipykernel
matplotlib
tensorboard
scikit-learn
torch>=2.0